refactor(npu): isolate DeepSeek V4 implementation
