+++ afd_plugin/model_executor/models/npu/deepseek_v4_attention_gate.py
@@ -9,7 +9,7 @@ from typing import TYPE_CHECKING
 import torch
 
 if TYPE_CHECKING:
-    from afd_plugin.model_executor.models.deepseek_v4 import (
+    from afd_plugin.model_executor.models.npu.deepseek_v4 import (
         AFDDeepseekV4AttentionGateRemoteMoE,
     )
